fix egghead build
