+++ apps/egghead/src/lib/posts-query.ts
@@ -9,23 +9,18 @@ import {
 	contentResource,
 	contentResourceResource,
 	contentResourceTag as contentResourceTagTable,
+	contentResourceVersion as contentResourceVersionTable,
 	contributionTypes,
 	users,
 } from '@/db/schema'
 import {
-	createNewPostVersion,
-	deletePostFromDatabase,
 	generateContentHash,
-	getVideoDuration,
 	NewPost,
 	Post,
 	PostAction,
 	PostSchema,
-	PostState,
 	PostUpdate,
-	PostVisibility,
-	writeNewPostToDatabase,
-	writePostUpdateToDatabase,
+	updatePostSlug,
 } from '@/lib/posts'
 import { getServerAuthSession } from '@/server/auth'
 import { guid } from '@/utils/guid'
@@ -41,12 +36,10 @@ import 'server-only'
 import { getMuxAsset } from '@coursebuilder/core/lib/mux'
 
 import {
+	crreateEggheadLesson,
 	determineEggheadLessonState,
 	determineEggheadVisibilityState,
-	EggheadLessonState,
-	EggheadLessonVisibilityState,
 	getEggheadUserProfile,
-	removeLegacyTaggingsOnEgghead,
 	updateEggheadLesson,
 	writeLegacyTaggingsToEgghead,
 } from './egghead'
@@ -282,4 +275,244 @@ export async function removeTagFromPost(postId: string, tagId: string) {
 			),
 		)
 	await writeLegacyTaggingsToEgghead(postId)
+}
+
+export async function writeNewPostToDatabase(input: {
+	newPost: NewPost
+	eggheadInstructorId: number
+	createdById: string
+}) {
+	const { title, videoResourceId } = input.newPost
+	const { eggheadInstructorId, createdById } = input
+	const postGuid = guid()
+	const newPostId = `post_${postGuid}`
+	const videoResource =
+		await courseBuilderAdapter.getVideoResource(videoResourceId)
+
+	const eggheadLessonId = await crreateEggheadLesson({
+		title: title,
+		slug: `${slugify(title)}~${postGuid}`,
+		instructorId: eggheadInstructorId,
+	})
+
+	await db
+		.insert(contentResource)
+		.values({
+			id: newPostId,
+			type: 'post',
+			createdById,
+			fields: {
+				title,
+				state: 'draft',
+				visibility: 'unlisted',
+				slug: `${slugify(title)}~${postGuid}`,
+				eggheadLessonId,
+			},
+		})
+		.catch((error) => {
+			console.error('🚨 Error creating post', error)
+			throw error
+		})
+
+	const post = await getPost(newPostId)
+
+	await createNewPostVersion(post)
+
+	if (post) {
+		if (videoResource) {
+			await db
+				.insert(contentResourceResource)
+				.values({ resourceOfId: post.id, resourceId: videoResource.id })
+		}
+
+		const contributionType = await db.query.contributionTypes.findFirst({
+			where: eq(contributionTypes.slug, 'author'),
+		})
+
+		if (contributionType) {
+			await db.insert(contentContributions).values({
+				id: `cc-${guid()}`,
+				userId: createdById,
+				contentId: post.id,
+				contributionTypeId: contributionType.id,
+			})
+		}
+
+		return post
+	}
+	return null
+}
+
+export async function writePostUpdateToDatabase(input: {
+	currentPost: Post
+	postUpdate: PostUpdate
+	action: PostAction
+	updatedById: string
+}) {
+	const {
+		currentPost = await getPost(input.postUpdate.id),
+		postUpdate,
+		action = 'save',
+		updatedById,
+	} = input
+
+	if (!currentPost) {
+		throw new Error(`Post with id ${input.postUpdate.id} not found.`)
+	}
+
+	if (!postUpdate.fields.title) {
+		throw new Error('Title is required')
+	}
+
+	let postSlug = updatePostSlug(currentPost, postUpdate.fields.title)
+
+	if (postUpdate.fields.title !== currentPost.fields.title) {
+		const splitSlug = currentPost?.fields.slug.split('~') || ['', guid()]
+		postSlug = `${slugify(postUpdate.fields.title ?? '')}~${splitSlug[1] || guid()}`
+	}
+
+	const lessonState = determineEggheadLessonState(
+		action,
+		postUpdate.fields.state,
+	)
+	const lessonVisibilityState = determineEggheadVisibilityState(
+		postUpdate.fields.visibility,
+		postUpdate.fields.state,
+	)
+
+	const duration = await getVideoDuration(currentPost.resources)
+	const timeToRead = Math.floor(
+		readingTime(currentPost.fields.body ?? '').time / 1000,
+	)
+
+	if (currentPost.fields.eggheadLessonId) {
+		await updateEggheadLesson({
+			eggheadLessonId: currentPost.fields.eggheadLessonId,
+			state: lessonState,
+			visibilityState: lessonVisibilityState,
+			duration: duration > 0 ? duration : timeToRead,
+		})
+	}
+
+	await courseBuilderAdapter.updateContentResourceFields({
+		id: currentPost.id,
+		fields: {
+			...currentPost.fields,
+			...postUpdate.fields,
+			slug: postSlug,
+		},
+	})
+
+	const updatedPost = await getPost(currentPost.id)
+
+	if (!updatedPost) {
+		throw new Error(`Post with id ${currentPost.id} not found.`)
+	}
+
+	const newContentHash = generateContentHash(updatedPost)
+	const currentContentHash = currentPost.currentVersionId?.split('~')[1]
+
+	if (newContentHash !== currentContentHash) {
+		await createNewPostVersion(updatedPost, updatedById)
+	}
+
+	await upsertPostToTypeSense(updatedPost)
+
+	return updatedPost
+}
+
+export async function deletePostFromDatabase(id: string) {
+	const post = PostSchema.nullish().parse(
+		await db.query.contentResource.findFirst({
+			where: eq(contentResource.id, id),
+			with: {
+				resources: true,
+			},
+		}),
+	)
+
+	if (!post) {
+		throw new Error(`Post with id ${id} not found.`)
+	}
+
+	if (post.fields.eggheadLessonId) {
+		await eggheadPgQuery(
+			`UPDATE lessons SET state = 'retired' WHERE id = ${post.fields.eggheadLessonId}`,
+		)
+	}
+
+	await db
+		.delete(contentResourceResource)
+		.where(eq(contentResourceResource.resourceOfId, id))
+
+	await db.delete(contentResource).where(eq(contentResource.id, id))
+
+	return true
+}
+
+export async function createNewPostVersion(
+	post: Post | null,
+	createdById?: string,
+) {
+	if (!post) return null
+	const contentHash = generateContentHash(post)
+	const versionId = `version~${contentHash}`
+
+	// Check if this version already exists
+	const existingVersion = await db.query.contentResourceVersion.findFirst({
+		where: eq(contentResourceVersionTable.id, versionId),
+	})
+
+	if (existingVersion) {
+		// If this exact version already exists, just update the current version pointer
+		await db
+			.update(contentResource)
+			.set({ currentVersionId: versionId })
+			.where(eq(contentResource.id, post.id))
+		return post
+	}
+
+	// If it's a new version, create it
+	await db.transaction(async (trx) => {
+		await trx.insert(contentResourceVersionTable).values({
+			id: versionId,
+			resourceId: post.id,
+			parentVersionId: post.currentVersionId,
+			versionNumber: (await getLatestVersionNumber(post.id)) + 1,
+			fields: {
+				...post.fields,
+			},
+			createdAt: new Date(),
+			createdById: createdById ? createdById : post.createdById,
+		})
+
+		post.currentVersionId = versionId
+
+		await trx
+			.update(contentResource)
+			.set({ currentVersionId: versionId })
+			.where(eq(contentResource.id, post.id))
+	})
+	return post
+}
+
+export async function getLatestVersionNumber(postId: string): Promise<number> {
+	const latestVersion = await db.query.contentResourceVersion.findFirst({
+		where: eq(contentResourceVersionTable.resourceId, postId),
+		orderBy: desc(contentResourceVersionTable.versionNumber),
+	})
+	return latestVersion ? latestVersion.versionNumber : 0
+}
+
+export async function getVideoDuration(
+	resources: Post['resources'],
+): Promise<number> {
+	const videoResource = resources?.find(
+		(resource) => resource.resource.type === 'videoResource',
+	)
+	if (videoResource) {
+		const muxAsset = await getMuxAsset(videoResource.resource.fields.muxAssetId)
+		return muxAsset?.duration ? Math.floor(muxAsset.duration) : 0
+	}
+	return 0
 }
